+++ Calculator/Calculator/CalculatorItemQueue.swift
@@ -15,21 +15,7 @@ extension Int: CalculateItem {
     
 }
 
-class Node<T> {
-    var data: T
-    var next: Node?
-    
-    init(data: T, next: Node? = nil) {
-        self.data = data
-        self.next = next
-    }
-}
 
-class LinkdeList<T> {
-    var head: Node<T>?
-    var tail: Node<T>?
-
-}
 
 class CalculatorItemQueue<T: CalculateItem> {
     let linkdeList = LinkdeList<T>()
